Add ChMeetings export button to CSVExport.php with admin API integration
- Add ChMeetings Export section to CSVExport.php UI with download button
- Implement fetch-based API call to /admin/api/database/people/export/chmeetings
- Auto-trigger file download with date-stamped filename (ChMeetings-YYYY-MM-DD.csv)
- Add loading state feedback and error/success notifications via window.CRM.notify()
- Use fetch() for file downloads since CSV endpoints return raw text, not JSON
- All UI text wrapped with gettext() for internationalization

This allows admins to easily export all people data in ChMeetings format directly from the CSV Export page.
