+++ src/admin/routes/api/database.php
@@ -2,13 +2,18 @@
 
 use ChurchCRM\Authentication\AuthenticationManager;
 use ChurchCRM\Slim\SlimUtils;
+use ChurchCRM\Utils\CsvExporter;
 use ChurchCRM\Utils\LoggerUtils;
 use Propel\Runtime\Propel;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Slim\Routing\RouteCollectorProxy;
+use ChurchCRM\dto\SystemConfig;
+use ChurchCRM\model\ChurchCRM\PersonQuery;
 
 $app->group('/api/database', function (RouteCollectorProxy $group): void {
+
+    $group->get('/people/export/chmeetings', 'exportChMeetings');
 
     /**
      * Reset database - drops all tables and views
@@ -86,3 +91,92 @@ $app->group('/api/database', function (RouteCollectorProxy $group): void {
     });
 
 });
+
+function exportChMeetings(Request $request, Response $response, array $args): Response
+{
+    $header_data = [
+        'First Name',
+        'Last Name',
+        'Middle Name',
+        'Gender',
+        'Marital Status',
+        'Anniversary',
+        'Engagement Date',
+        'Birthdate',
+        'Mobile Phone',
+        'Home Phone',
+        'Email',
+        'Facebook',
+        'School',
+        'Grade',
+        'Employer',
+        'Job Title',
+        'Talents And Hobbies',
+        'Address Line',
+        'Address Line 2',
+        'City',
+        'State',
+        'ZIP Code',
+        'Notes',
+        'Join Date',
+        'Family Id',
+        'Family Role',
+        'Baptism Date',
+        'Baptism Location',
+        'Nickname',
+    ];
+    $people = PersonQuery::create()->find();
+    $list = [];
+    foreach ($people as $person) {
+        $family = $person->getFamily();
+        $anniversary = ($family ? $family->getWeddingdate(SystemConfig::getValue('sDateFormatLong')) : '');
+        $familyRole = $person->getFamilyRoleName();
+        if ($familyRole === 'Head of Household') {
+            $familyRole = 'Primary';
+        }
+
+        $chPerson = [
+            $person->getFirstName(),
+            $person->getLastName(),
+            $person->getMiddleName(),
+            $person->getGenderName(),
+            '',
+            $anniversary,
+            '',
+            $person->getFormattedBirthDate(),
+            $person->getCellPhone(),
+            $person->getHomePhone(),
+            $person->getEmail(),
+            $person->getFacebook(),
+            '',
+            '',
+            '',
+            '',
+            '',
+            $person->getAddress1(),
+            $person->getAddress2(),
+            $person->getCity(),
+            $person->getState(),
+            $person->getZip(),
+            '',
+            $person->getMembershipDate(SystemConfig::getValue('sDateFormatLong')),
+            $family ? $family->getId() : '',
+            $familyRole,
+            '',
+            '',
+            ''
+        ];
+        $list[] = $chPerson;
+    }
+
+    // Use CsvExporter for RFC 4180 compliance and formula injection prevention
+    CsvExporter::create(
+        $header_data,
+        $list,
+        'ChMeetings',
+        SystemConfig::getValue('sCSVExportCharset'),
+        true
+    );
+
+    return $response;
+}
