+++ src/api/routes/system/system-database.php
@@ -5,7 +5,6 @@ use ChurchCRM\Backup\BackupDownloader;
 use ChurchCRM\Backup\BackupJob;
 use ChurchCRM\Backup\RestoreJob;
 use ChurchCRM\dto\SystemConfig;
-use ChurchCRM\model\ChurchCRM\PersonQuery;
 use ChurchCRM\Slim\Middleware\Request\Auth\AdminRoleAuthMiddleware;
 use ChurchCRM\Slim\SlimUtils;
 use ChurchCRM\Utils\LoggerUtils;
@@ -15,8 +14,6 @@ use Psr\Http\Message\ServerRequestInterface as Request;
 use Slim\Routing\RouteCollectorProxy;
 
 $app->group('/database', function (RouteCollectorProxy $group): void {
-
-    $group->get('/people/export/chmeetings', 'exportChMeetings');
 
     $group->post('/backup', function (Request $request, Response $response, array $args): Response {
         $input = $request->getParsedBody();
@@ -95,99 +92,3 @@ $app->group('/database', function (RouteCollectorProxy $group): void {
         return $response;
     });
 })->add(AdminRoleAuthMiddleware::class);
-
-function exportChMeetings(Request $request, Response $response, array $args): Response
-{
-    $header_data = [
-        'First Name',
-        'Last Name',
-        'Middle Name',
-        'Gender',
-        'Marital Status',
-        'Anniversary',
-        'Engagement Date',
-        'Birthdate',
-        'Mobile Phone',
-        'Home Phone',
-        'Email',
-        'Facebook',
-        'School',
-        'Grade',
-        'Employer',
-        'Job Title',
-        'Talents And Hobbies',
-        'Address Line',
-        'Address Line 2',
-        'City',
-        'State',
-        'ZIP Code',
-        'Notes',
-        'Join Date',
-        'Family Id',
-        'Family Role',
-        'Baptism Date',
-        'Baptism Location',
-        'Nickname',
-    ];
-    $people = PersonQuery::create()->find();
-    $list = [];
-    foreach ($people as $person) {
-        $family = $person->getFamily();
-        $anniversary = ($family ? $family->getWeddingdate(SystemConfig::getValue('sDateFormatLong')) : '');
-        $familyRole = $person->getFamilyRoleName();
-        if ($familyRole === 'Head of Household') {
-            $familyRole = 'Primary';
-        }
-
-        $chPerson = [
-            $person->getFirstName(),
-            $person->getLastName(),
-            $person->getMiddleName(),
-            $person->getGenderName(),
-            '',
-            $anniversary,
-            '',
-            $person->getFormattedBirthDate(),
-            $person->getCellPhone(),
-            $person->getHomePhone(),
-            $person->getEmail(),
-            $person->getFacebook(),
-            '',
-            '',
-            '',
-            '',
-            '',
-            $person->getAddress1(),
-            $person->getAddress2(),
-            $person->getCity(),
-            $person->getState(),
-            $person->getZip(),
-            '',
-            $person->getMembershipDate(SystemConfig::getValue('sDateFormatLong')),
-            $family ? $family->getId() : '',
-            $familyRole,
-            '',
-            '',
-            ''
-        ];
-        $list[] = $chPerson;
-    }
-
-    $out = fopen('php://temp', 'w+');
-    fputcsv($out, $header_data);
-    foreach ($list as $fields) {
-        fputcsv($out, $fields, ',');
-    }
-    rewind($out);
-    $csvData = stream_get_contents($out);
-    fclose($out);
-
-    $response = $response->withHeader('Content-Type', 'text/csv');
-    $response = $response->withHeader(
-        'Content-Disposition',
-        'attachment; filename="ChMeetings-' . date(SystemConfig::getValue('sDateFilenameFormat')) . '.csv"'
-    );
-    $response->getBody()->write($csvData);
-
-    return $response;
-}
